+++ packages/typespec-ts/src/modular/helpers/namingHelpers.ts
@@ -27,9 +27,10 @@ export interface GuardedName {
 }
 
 export function getOperationName(operation: ServiceOperation): GuardedName {
-  if (isReservedName(operation.name, NameType.Operation)) {
+  const norm = normalizeName(operation.name, NameType.Method, true);
+  if (isReservedName(operation.name, NameType.Method)) {
     return {
-      name: `$${operation.name}`,
+      name: norm,
       fixme: [
         `${operation.name} is a reserved word that cannot be used as an operation name. 
         Please add @clientName("clientName") or @clientName("<JS-Specific-Name>", "javascript") 
@@ -37,9 +38,8 @@ export function getOperationName(operation: ServiceOperation): GuardedName {
       ]
     };
   }
-
   return {
-    name: normalizeName(operation.name, NameType.Operation, true)
+    name: norm
   };
 }